Move InvokerState to common
